// types/tmdb/Movie.ts
export interface Movie {
  adult: boolean;
  backdrop_path: string | null;
  genre_ids: number[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string | null;
  popularity: number;
  poster_path: string | null;
  release_date?: string;
  title: string;
  video: boolean;
  vote_average?: number;
  vote_count: number;
}

export interface MovieWithMediaType extends Movie {
  media_type: "movie";
}